moved reformatting function to utils file for cleaner component
